requirements.txt
selenium
webdriver-manager
ddt
mysql-connector-python